+++ pkg/yamlprocessor/BUILD.bazel
@@ -2,8 +2,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["yamlutil.go"],
-    importpath = "github.com/pipe-cd/pipe/pkg/yamlutil",
+    srcs = ["yamlprocessor.go"],
+    importpath = "github.com/pipe-cd/pipe/pkg/yamlprocessor",
     visibility = ["//visibility:public"],
     deps = [
         "@com_github_goccy_go_yaml//:go_default_library",
@@ -15,7 +15,7 @@ go_library(
 go_test(
     name = "go_default_test",
     size = "small",
-    srcs = ["yamlutil_test.go"],
+    srcs = ["yamlprocessor_test.go"],
     embed = [":go_default_library"],
     deps = ["@com_github_stretchr_testify//assert:go_default_library"],
 )
